split the rest of gcp/api.py to individual files
