+++ tools/run_tests/xds_k8s_test_driver/framework/infrastructure/gcp/network_services.py
@@ -19,11 +19,12 @@ from google.rpc import code_pb2
 import tenacity
 
 from framework.infrastructure import gcp
+from framework.infrastructure.gcp._internal import gcp_api_standard as _gcp_api_standard
 
 logger = logging.getLogger(__name__)
 
 
-class NetworkServicesV1Alpha1(gcp.api.GcpStandardCloudApiResource):
+class NetworkServicesV1Alpha1(_gcp_api_standard.GcpApiStandard):
     ENDPOINT_CONFIG_SELECTORS = 'endpointConfigSelectors'
 
     @dataclasses.dataclass(frozen=True)
@@ -92,5 +93,5 @@ class NetworkServicesV1Alpha1(gcp.api.GcpStandardCloudApiResource):
 
     @staticmethod
     def _operation_internal_error(exception):
-        return (isinstance(exception, gcp.api.OperationError) and
+        return (isinstance(exception, gcp.GcpOperationError) and
                 exception.error.code == code_pb2.INTERNAL)
